@moredure 2022-03-20
Also bufpool.Put will allocate byte slice escapes to the heap...