@PeteBassettBet365 2022-03-22
> Also bufpool.Put will allocate byte slice escapes to the heap...

Hi. The original slices in the tdsBuffer were also on the heap, but they were garbage collected, we now reuse them due to the pool. 

Is there any other feedback about this PR?